fix action not define when add to wishlist

+++ organi/views/wishlist.php
@@ -24,37 +24,38 @@
 ?>
 
 <?php
-if ($_GET['action'] === 'add') {
-    // Kiểm tra xem sản phẩm đã tồn tại trong danh sách yêu thích của người dùng hay chưa
-    $productId = $_GET['id'];
-    $result = $connect->query("SELECT * FROM `products` WHERE `id` = $productId");
-    $result = mysqli_fetch_array($result);
-    $member = $_SESSION['member'];
-    $memberId = $connect->query("SELECT `id` FROM `member` WHERE `username` = '$member'");
-    $memberId = mysqli_fetch_array($memberId);
-    $memberId = $memberId['id'];
-    $checkQuery = "SELECT * FROM `wishlist` WHERE `member_id` = '$memberId' AND `product_id` = $productId";
-    $checkStatement = $connect->query($checkQuery);
-    if (mysqli_num_rows($checkStatement) != 0) {
-        confirm('Đã có sản phẩm này trong wishlist');
-        header("location: ?option=wishlist");
-    } else {
-        // Nếu sản phẩm chưa tồn tại, thêm vào danh sách yêu thích
-        $productName = $result['name'];
-        $productPrice = $result['price'];
-        $productImg = $result['image'];
-        $insertQuery = "INSERT INTO `wishlist` (`member_id`, `product_id`, `product_name`, `product_image`, `product_price`) 
-                        VALUES ($memberId, $productId, '$productName', '$productImg', $productPrice)";
-        $insertStatement = $connect->query($insertQuery);
+if (isset($_GET['action'])) {
+    if ($_GET['action'] === 'add') {
+        // Kiểm tra xem sản phẩm đã tồn tại trong danh sách yêu thích của người dùng hay chưa
+        $productId = $_GET['id'];
+        $result = $connect->query("SELECT * FROM `products` WHERE `id` = $productId");
+        $result = mysqli_fetch_array($result);
+        $member = $_SESSION['member'];
+        $memberId = $connect->query("SELECT `id` FROM `member` WHERE `username` = '$member'");
+        $memberId = mysqli_fetch_array($memberId);
+        $memberId = $memberId['id'];
+        $checkQuery = "SELECT * FROM `wishlist` WHERE `member_id` = '$memberId' AND `product_id` = $productId";
+        $checkStatement = $connect->query($checkQuery);
+        if (mysqli_num_rows($checkStatement) != 0) {
+            //confirm('Đã có sản phẩm này trong wishlist');
+            header("location: ?option=wishlist");
+        } else {
+            // Nếu sản phẩm chưa tồn tại, thêm vào danh sách yêu thích
+            $productName = $result['name'];
+            $productPrice = $result['price'];
+            $productImg = $result['image'];
+            $insertQuery = "INSERT INTO `wishlist` (`member_id`, `product_id`, `product_name`, `product_image`, `product_price`) 
+                            VALUES ($memberId, $productId, '$productName', '$productImg', $productPrice)";
+            $insertStatement = $connect->query($insertQuery);
+            header("location: ?option=wishlist");
+        }
+    } else if ($_GET['action'] === 'remove') {
+        $Id = $_GET['id'];
+        $deleteQuery = "DELETE FROM `wishlist` WHERE `id` = $Id";
+        $deleteStatement = $connect->query($deleteQuery);
         header("location: ?option=wishlist");
     }
-} else if ($_GET['action'] === 'remove') {
-    $Id = $_GET['id'];
-    $deleteQuery = "DELETE FROM `wishlist` WHERE `id` = $Id";
-    $deleteStatement = $connect->query($deleteQuery);
-    header("location: ?option=wishlist");
 }
-
 ?>
 
 
@@ -87,53 +88,52 @@ if ($_GET['action'] === 'add') {
                 <?php
                 // Kiểm tra xem người dùng đã đăng nhập hay chưa
                 if (isset($_SESSION['member'])) {
-                    ?>
-                <div>
-                    <table class="wishlist__table">
-                        <?php
-                                $member = $_SESSION['member'];
-                                $memberId = $connect->query("SELECT `id` FROM `member` WHERE `username` = '$member'");
-                                $memberId = mysqli_fetch_array($memberId);
-                                $memberId = $memberId['id'];
-                                $query = "SELECT * FROM `wishlist` WHERE member_id IN ($memberId)";
-                                $result = $connect->query($query);
-                                if (mysqli_num_rows($result) != 0) { 
+                ?>
+                    <div>
+                        <table class="wishlist__table">
+                            <?php
+                            $member = $_SESSION['member'];
+                            $memberId = $connect->query("SELECT `id` FROM `member` WHERE `username` = '$member'");
+                            $memberId = mysqli_fetch_array($memberId);
+                            $memberId = $memberId['id'];
+                            $query = "SELECT * FROM `wishlist` WHERE member_id IN ($memberId)";
+                            $result = $connect->query($query);
+                            if (mysqli_num_rows($result) != 0) {
                             ?>
-                        <thead>
-                            <tr>
-                                <th class="wishlist__product">Tên Sách</th>
-                                <th>Giá</th>
-                                <th>Xóa</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <?php
+                                <thead>
+                                    <tr>
+                                        <th class="wishlist__product">Tên Sách</th>
+                                        <th>Giá</th>
+                                        <th>Xóa</th>
+                                    </tr>
+                                </thead>
+                                <tbody>
+                                    <?php
                                     foreach ($result as $item) :
-                                ?>
-                            <tr>
-                                <td class="wishlist__item">
-                                    <img width="100px" src="../images/<?= $item['product_image'] ?>" alt="">
-                                    <h5>
-                                        <?= $item['product_name'] ?>
-                                    </h5>
-                                </td>
-                                <td class="wishlist__price">
-                                    <?= number_format($item['product_price'], 0, ',', '.') ?>đ
-                                </td>
-                                <td class="wishlist__remove">
-                                    <span class="icon_close"
-                                        onclick="if (confirm('Bạn có chắc chắn muốn xóa sản phẩm khỏi wishlist?')) location='?option=wishlist&action=remove&id=<?= $item['id'] ?>';"></span>
-                                </td>
-                            </tr>
-                            <?php
-                                endforeach;
+                                    ?>
+                                        <tr>
+                                            <td class="wishlist__item">
+                                                <img width="100px" src="../images/<?= $item['product_image'] ?>" alt="">
+                                                <h5>
+                                                    <?= $item['product_name'] ?>
+                                                </h5>
+                                            </td>
+                                            <td class="wishlist__price">
+                                                <?= number_format($item['product_price'], 0, ',', '.') ?>đ
+                                            </td>
+                                            <td class="wishlist__remove">
+                                                <span class="icon_close" onclick="if (confirm('Bạn có chắc chắn muốn xóa sản phẩm khỏi wishlist?')) location='?option=wishlist&action=remove&id=<?= $item['id'] ?>';"></span>
+                                            </td>
+                                        </tr>
+                                <?php
+                                    endforeach;
                                 } else {
                                     echo '<h3 style="text-align: center;">Danh sách yêu thích trống</h3></td></tr>';
                                 }
                                 ?>
-                        </tbody>
-                    </table>
-                </div>
+                                </tbody>
+                        </table>
+                    </div>
                 <?php
                 } else {
                     // Hiển thị thông báo nếu chưa đăng nhập
